% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{draw_a_word}
\alias{draw_a_word}
\title{Draw a single word given topic and word distributions}
\usage{
draw_a_word(theta, phi)
}
\arguments{
\item{theta}{Topic distribution, length \eqn{k} vector}

\item{phi}{Word distribution for all topics, \eqn{k \times v} matrix}
}
\value{
Integer representation for a single word
}
\description{
Draw a single word given topic and word distributions
}
