% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{draw_words}
\alias{draw_words}
\title{Draw words for one document}
\usage{
draw_words(Ni, theta, phi)
}
\arguments{
\item{Ni}{length of document i}

\item{theta}{\code{theta_i}, topic distribution for document i}

\item{phi}{word distribution for all topics}
}
\value{
Document-term matrix (single document), as a tibble, with columns \code{word} and \code{n}
}
\description{
Draw words for one document
}
