% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellinger.R
\name{hellinger.data.frame}
\alias{hellinger.data.frame}
\title{Hellinger distance for dataframes}
\usage{
\method{hellinger}{data.frame}(
  topics1,
  id1 = "document",
  cat1 = "topic",
  prob1 = "prob",
  topics2 = NULL,
  id2 = "document",
  cat2 = "topic",
  prob2 = "prob",
  df = FALSE,
  ...
)
}
\arguments{
\item{topics1}{First tidied topic model dataframe}

\item{id1}{Unit identifier column in \code{topics1} (DOIs, auids, ORU name, etc.)}

\item{cat1}{Category identifier column in \code{topics1} (topics)}

\item{prob1}{Probability value column in \code{topics1} (gamma)}

\item{topics2}{Optional second tidied topic model dataframe}

\item{id2}{Unit identifier column in \code{topics2}}

\item{cat2}{Category identifier column in \code{topics2}}

\item{prob2}{Probability value column in \code{topics2}}

\item{df}{Should the function return the matrix of Hellinger distances (default) or a tidy dataframe?}

\item{...}{Not used; required for S3 method compatibility}
}
\value{
matrix or tidy dataframe (default) of Hellinger distances
}
\description{
Hellinger distances, either pairwise within a single tidied topic model dataframe or between two tidied topic model dataframes
}
\examples{
set.seed(2022-06-09)
topics1 = rdirichlet(3, rep(5, 5)) |>
    tibble::as_tibble(rownames = 'doc_id',
                      .name_repair = tmfast:::make_colnames) |>
    dplyr::mutate(doc_id = stringr::str_c('doc_', doc_id)) |>
    tidyr::pivot_longer(tidyselect::starts_with('V'),
                        names_to = 'topic',
                        values_to = 'gamma')
topics2 = rdirichlet(3, rep(5, 5)) |>
    tibble::as_tibble(rownames = 'doc_id',
                      .name_repair = tmfast:::make_colnames) |>
    dplyr::mutate(doc_id = stringr::str_c('doc_', as.integer(doc_id) + 5)) |>
    tidyr::pivot_longer(tidyselect::starts_with('V'),
                        names_to = 'topic',
                        values_to = 'gamma')
hellinger(topics1, doc_id, prob1 = 'gamma', df = TRUE)
hellinger(topics1, doc_id, prob1 = 'gamma',
          topics2 = topics2, id2 = doc_id, prob2 = 'gamma')
}
