% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellinger.R
\name{hellinger.Matrix}
\alias{hellinger.Matrix}
\title{Hellinger distance for matrices}
\usage{
\method{hellinger}{Matrix}(topics1, topics2 = NULL, ...)
}
\arguments{
\item{topics1}{First matrix, \eqn{n_1 \times k}}

\item{topics2}{Optional second matrix, \eqn{n_2 \times k}}

\item{...}{Not used; required for S3 method compatibility}
}
\value{
Matrix of size \eqn{n_1 \times n_1} or \eqn{n_1 \times n_2}
}
\description{
Calculates Hellinger distance for each pair of rows in the given matrix, or each combination of rows from the two matrices
}
\examples{
set.seed(2022-06-09)
topics1 = rdirichlet(3, rep(5, 5))
topics2 = rdirichlet(3, rep(5, 5))
hellinger(topics1)
hellinger(topics1, topics2)
}
