% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellinger.R
\name{hellinger}
\alias{hellinger}
\title{Hellinger distances}
\usage{
hellinger(topics1, ...)
}
\arguments{
\item{topics1}{Object to dispatch on}

\item{...}{Passed to methods}
}
\description{
Hellinger distances
}
