% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{journal_specific}
\alias{journal_specific}
\title{"Journal-specific" simulation scenario}
\usage{
journal_specific(
  k = 5,
  Mj = 100,
  topic_peak = 0.8,
  topic_scale = 10,
  word_beta = 0.01,
  vocab = 10 * Mj * k,
  size = 3,
  mu = 300,
  bigjournal = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{k}{Number of topics/journals}

\item{Mj}{Number of documents from each journal}

\item{topic_peak}{Peak value for the asymmetric Dirichlet prior for true topic-doc distributions}

\item{topic_scale}{Scale for the asymmetric Dirichlet prior for true topic-doc distributions}

\item{word_beta}{Parameter for the symmetric Dirichlet prior for true word-doc distributions}

\item{vocab}{Size of the vocabulary}

\item{size}{Size parameter for the negative binomial distribution of document lengths}

\item{mu}{Mean parameter for the negative binomial distribution of document lengths}

\item{bigjournal}{Should the first journal have documents 10x as long (on average) as the others?}

\item{verbose}{When TRUE, sends messages about the progress of the simulation}
}
\value{
A one-row \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\describe{
\item{phi}{Mean Hellinger distance between true and fitted word-topic distributions}
\item{phi_vec}{List-column of per-topic Hellinger distances}
\item{theta}{Mean Hellinger distance between true and fitted document-topic distributions}
\item{theta_vec}{List-column of per-document Hellinger distances}
}
}
\description{
Generates a corpus with \code{Mj} documents from \code{k} journals, each of which has a characteristic topic.  Fits a varimax topic model of rank \code{k}, rotates the word-topic distribution to align with the true values, and reports Hellinger distance comparisons for each topic (word-topic) and document (topic-doc).
}
\examples{
journal_specific(k = 2, Mj = 10, vocab = 50, verbose = FALSE)
}
\seealso{
Other generators: 
\code{\link{draw_corpus}()},
\code{\link{peak_alpha}()},
\code{\link{rdirichlet}()}
}
\concept{generators}
