% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{peak_alpha}
\alias{peak_alpha}
\title{Alpha parameter with a single peak}
\usage{
peak_alpha(k, i, peak = 0.8, scale = 1)
}
\arguments{
\item{k}{Number of components}

\item{i}{Index for the component that takes value \code{peak}}

\item{peak}{Value for the single peak component}

\item{scale}{Scaling factor applied to all concentration parameters}
}
\value{
Vector of length \code{k}
}
\description{
This function allows us to quickly define an \code{alpha} parameter for a Dirichlet distribution with a single (presumably high) \code{peak*scale} value at component \code{i} and all other components a uniform (presumably low) value \code{(1-peak)/(k-1)*scale}.
}
\examples{
peak_alpha(5, 2)
peak_alpha(5, 2, peak = 0.9, scale = 10)
}
\seealso{
Other generators: 
\code{\link{draw_corpus}()},
\code{\link{journal_specific}()},
\code{\link{rdirichlet}()}
}
\concept{generators}
