% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renormalize.R
\name{solve_power}
\alias{solve_power}
\title{Solve the equation to find the desired exponent}
\usage{
solve_power(p, target_H, return_full = FALSE)
}
\arguments{
\item{p}{Initial distribution}

\item{target_H}{Desired entropy for the transformed distribution}

\item{return_full}{Return the full uniroot() output?}
}
\value{
Numeric value of the desired exponent
}
\description{
After \url{https://stats.stackexchange.com/questions/521582/controlling-the-entropy-of-a-distribution}
}
\examples{
p = c(0.5, 0.3, 0.2)
solve_power(p, target_H = 1.0)
}
