% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{umap.matrix}
\alias{umap.matrix}
\title{Discursive space with UMAP given a distance matrix}
\usage{
\method{umap}{matrix}(dist_mx, include_data = FALSE, df = TRUE, ...)
}
\arguments{
\item{dist_mx}{Distance matrix}

\item{include_data}{By default, to save space the data (distance matrix) is not returned}

\item{df}{Return a tibble with columns \code{document}, \code{x}, and \code{y} (default) or the output of \code{umap}.}

\item{...}{Other parameters passed to \code{umap::umap()}}
}
\value{
Object of class \code{umap}, with components \code{layout} (coordinates of items), \code{knn} (k-nearest neighbors matrices), \code{config} (UMAP configuration) \emph{or} tidied dataframe, per argument \code{df}
}
\description{
Construct a 2-dimensional "discursive space" embedding given a distance matrix.
}
\examples{
gamma = rdirichlet(26, 1, 5)
rownames(gamma) = letters
h_gamma = hellinger(gamma)
embedded = umap(h_gamma, df = TRUE, verbose = TRUE)
}
