% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{umap.STM}
\alias{umap.STM}
\title{Discursive space with UMAP for structural topic models}
\usage{
\method{umap}{STM}(tm, doc_ids, ...)
}
\arguments{
\item{tm}{\code{STM} object}

\item{doc_ids}{Character vector of document IDs}

\item{...}{Other arguments, passed to \code{umap.matrix()}}
}
\value{
\code{umap} object or tidied dataframe; see \code{umap.matrix()} argument \code{df}
}
\description{
Discursive space with UMAP for structural topic models
}
