% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{draw_corpus}
\alias{draw_corpus}
\title{Draw a collection of documents}
\usage{
draw_corpus(N, theta, phi)
}
\arguments{
\item{N}{Length of documents}

\item{theta}{Topic distribution for all documents, \eqn{n \times k} matrix}

\item{phi}{Word distribution for all topics, \eqn{k \times v} matrix}
}
\value{
Document-term matrix, as a tibble, with columns \code{doc}, \code{word}, and \code{n}
}
\description{
Draw a collection of documents
}
\seealso{
Other generators: 
\code{\link{journal_specific}()},
\code{\link{peak_alpha}()},
\code{\link{rdirichlet}()}
}
\concept{generators}
