% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmfast.R
\name{fit_varimax}
\alias{fit_varimax}
\title{Given a (rank \code{n}) PCA fit, return a rank \code{k < n} varimax fit}
\usage{
fit_varimax(k, pca, feature_names, obs_names, varimax_fn, varimax_opts)
}
\arguments{
\item{k}{Desired rank of the fitted varimax model}

\item{pca}{Fitted PCA model}

\item{feature_names}{Names of the features (eg, data columns)}

\item{obs_names}{Names of the observations (eg, data rows)}

\item{varimax_fn}{Function to use for varimax rotation}

\item{varimax_opts}{Options passed to \code{varimax_fn}}
}
\value{
List with components
- \code{loadings}: Rotated feature loadings
- \code{rotmat}:  Rotation matrix
- \code{scores}:  Rotated observation scores
}
\description{
Given a (rank \code{n}) PCA fit, return a rank \code{k < n} varimax fit
}
\details{
After the initial rotation, factors with negative skew (left tails) are flipped
}
