% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellinger.R
\name{hellinger.data.frame}
\alias{hellinger.data.frame}
\title{Hellinger distance}
\usage{
\method{hellinger}{data.frame}(
  topicsdf1,
  id1,
  cat1 = "topic",
  prob1 = "prob",
  topicsdf2 = NULL,
  id2 = NULL,
  cat2 = "topic",
  prob2 = "prob",
  df = FALSE
)
}
\arguments{
\item{df}{Should the function return the matrix of Hellinger distances (default) or a tidy dataframe?}

\item{{topicsdf1, topicsdf2}}{Tidied topic model dataframes}

\item{{id1, id2}}{Unit identifiers (DOIs, auids, ORU name, etc.)}

\item{{cat1, cat2}}{Category identifiers (topics)}

\item{{prob1, prob2}}{Probability values (gamma)}
}
\value{
matrix or tidy dataframe (default) of Hellinger distances
}
\description{
Hellinger distances, either pairwise within a single tidied topic model dataframe or between two tidied topic model dataframes
}
\examples{
set.seed(2022-06-09)
topics1 = rdirichlet(3, rep(5, 5)) |>
    tibble::as_tibble(rownames = 'doc_id', .name_repair = make_colnames) |>
    tidyr::pivot_longer(starts_with('V'),
                 names_to = 'topic',
                 values_to = 'gamma')
topics2 = rdirichlet(3, rep(5, 5)) |>
    tibble::as_tibble(rownames = 'doc_id', .name_repair = make_colnames) |>
    tidyr::pivot_longer(starts_with('V'),
                 names_to = 'topic',
                 values_to = 'gamma')
hellinger(topics1, doc_id, prob1 = 'gamma', df = TRUE)
hellinger(topics1, doc_id, prob1 = 'gamma',
          topicsdf2 = topics2, id2 = doc_id, prob2 = 'gamma')
}
