% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellinger.R
\name{hellinger.Matrix}
\alias{hellinger.Matrix}
\title{Hellinger distance for matrices}
\usage{
\method{hellinger}{Matrix}(mx1, mx2 = NULL)
}
\arguments{
\item{mx1}{First matrix, \eqn{n_1 \times k}}

\item{mx2}{Optional second matrix, \eqn{n_2 \times k}}
}
\value{
Matrix of size \eqn{n_1 \times n_1} or \eqn{n_1 \times n_2}
}
\description{
Calculates Hellinger distance for each pair of rows in the given matrix, or each combination of rows from the two matrices
}
\examples{
set.seed(2022-06-09)
mx1 = rdirichlet(3, rep(5, 5))
mx2 = rdirichlet(3, rep(5, 5))
hellinger(mx1)
hellinger(mx1, mx2)
}
