% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{loadings}
\alias{loadings}
\title{Extract a PCA/varimax loadings matrix}
\usage{
loadings(x, ...)
}
\arguments{
\item{x}{Object to dispatch on}

\item{...}{Passed to methods}
}
\description{
Extract a PCA/varimax loadings matrix
}
