% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{scores}
\alias{scores}
\title{Extract item scores from a fitted PCA/varimax model}
\usage{
scores(x, ...)
}
\arguments{
\item{x}{Object to dispatch on}

\item{...}{Passed to methods}
}
\description{
Extract item scores from a fitted PCA/varimax model
}
