% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{tidy.tmfast}
\alias{tidy.tmfast}
\title{Extract beta and gamma matrices from \code{tmfast} objects}
\usage{
\method{tidy}{tmfast}(x, k, matrix = c("beta", "gamma"), df = TRUE, rotation = NULL)
}
\arguments{
\item{x}{\code{tmfast} object}

\item{k}{Index (number of topics/factors)}

\item{matrix}{Desired matrix, either word-topic (\code{beta}) or topic-doc distributions (\code{gamma})}

\item{df}{Return a long dataframe (default) or wide matrix?}

\item{rotation}{Optional rotation matrix; see details}
}
\value{
A long dataframe, with one row per word-topic or topic-doc combination. Column names depend on the value of \code{matrix}.
}
\description{
Extract beta and gamma matrices from \code{tmfast} objects
}
\details{
If \code{rotation} is not \code{NULL}, loadings/scores will be rotated.  This might be used to align the fitted topics with known true topics, as in the \code{journal_specific} simulation.  Loadings are left-multiplied by the given rotation, while scores are right-multiplied by the transpose of the given rotation.
}
