% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{tidy_all}
\alias{tidy_all}
\title{Extract gamma or beta matrices for all topics}
\usage{
tidy_all(x, matrix = "beta", ...)
}
\arguments{
\item{x}{\code{tmfast} object}

\item{matrix}{Desired matrix, \code{'beta'} or \code{'gamma'}}

\item{...}{Other arguments, passed to \code{tidy.tmfast()}}
}
\value{
A long dataframe, with one row per word-topic or topic-doc combination. Column names depend on the value of \code{matrix}.
}
\description{
Extract gamma or beta matrices for all topics
}
\examples{
\donttest{
set.seed(42)
theta  = rdirichlet(50, 1, k = 4)
phi    = rdirichlet(4, 0.1, k = 20)
corpus = draw_corpus(rep(50L, 50), theta, phi)
model  = tmfast(corpus, n = c(3, 4))
tidy_all(model, matrix = 'beta')
}
}
