% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmfast.R
\docType{package}
\name{tmfast-package}
\alias{tmfast-package}
\title{Fitting "topic models" with PCA+varimax}
\description{
Fits topic models using varimax-rotated principal component analysis (PCA), following the "vintage factor analysis" approach of Rohe & Zheng (2020) \doi{10.48550/arXiv.2004.05387}. Leverages truncated PCA via 'irlba' for sparse matrices, enabling fast model fitting on large corpora. Includes an information-theoretic approach to vocabulary selection, 'broom'-compatible tidiers for extracting word-topic and topic-document matrices into a tidy data workflow, and samplers for constructing simulated corpora for benchmarking and method evaluation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dhicks.github.io/tmfast/}
  \item \url{https://github.com/dhicks/tmfast}
  \item Report bugs at \url{https://github.com/dhicks/tmfast/issues}
}

}
\author{
\strong{Maintainer}: D. Hicks \email{hicks.daniel.j@gmail.com} (\href{https://orcid.org/0000-0001-7945-4416}{ORCID}) [copyright holder]

}
