% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmfast.R
\docType{package}
\name{tmfast-package}
\alias{tmfast-package}
\title{Fitting "topic models" with PCA+varimax}
\description{
Fit topic models fast using varimax-rotated PCA
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dhicks.github.io/tmfast/}
  \item \url{https://github.com/dhicks/tmfast}
}

}
\author{
\strong{Maintainer}: Dan Hicks \email{hicks.daniel.j@gmail.com} (\href{https://orcid.org/0000-0001-7945-4416}{ORCID})

}
