% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmfast.R
\name{tmfast}
\alias{tmfast}
\title{Fit a topic model using PCA+varimax}
\usage{
tmfast(dtm, n, row = doc, column = word, value = n, verbose = FALSE, ...)
}
\arguments{
\item{dtm}{Document-term matrix.  Either an object inheriting from \code{Matrix} or a long dataframe representation with row column \code{row}, column column \code{column}, and value column \code{n}.}

\item{n}{Number of topics to return}

\item{row}{In dataframe \code{dtm}, row column}

\item{column}{In dataframe \code{dtm}, column column}

\item{value}{In dataframe \code{dtm}, value column}

\item{verbose}{Should \code{irlba()} be \code{verbose}?}

\item{...}{Other arguments, passed to \code{varimax_irlba}}
}
\value{
As per \code{varimax_irlba}, of class \code{tmfast}
}
\description{
Fit a topic model using PCA+varimax
}
\details{
If \code{dtm} is not a matrix, will be cast to a sparse matrix using \code{tidytext::case_sparse()}
}
