% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellinger.R
\name{tsne}
\alias{tsne}
\title{Discursive space}
\usage{
tsne(x, ...)
}
\arguments{
\item{tm}{A \code{tmfast} fitted topic model}

\item{k}{Number of topics}

\item{perplexity}{Perplexity parameter for t-SNE. By default, minimum of 30 and \code{floor((ndocs - 1)/3) - 1}.}

\item{df}{Return a dataframe with columns \code{doc}, \code{x}, and \code{y} (default) or the output of \code{Rtsne}.}
}
\value{
See \code{df}
}
\description{
2-dimensional "discursive space" representation of relationships between documents using Hellinger distances and t-SNE
}
\details{
Algorithm checks distances to 3*perplexity nearest neighbors.  Rtsne loses rownames (document IDs). Use \code{set.seed()} for reproducibility.
}
\examples{
## From the real books vignette
set.seed(42)
tsne(fitted_tmf, k = 4, df = TRUE) |>
    left_join(meta, by = c('doc' = 'book')) |>
    ggplot(aes(x, y, color = author)) +
    geom_point()
}
