% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{umap}
\alias{umap}
\title{Discursive space using UMAP}
\usage{
umap(x, ...)
}
\arguments{
\item{x}{Object to dispatch on}

\item{...}{Passed to methods}
}
\description{
2-dimensional "discursive space" representation of relationships between documents using Hellinger distances and UMAP
}
